requirements.txt
accelerate
fal-client
replicate
more-itertools
numpy
openai-whisper
opencc-python-reimplemented
optimum
tiktoken
torch
tqdm
transformers
